Modified:   Defined a let variable `disabledItems` with a value of an empty array (src/App.svelte)             Changed a function `handleRemoveToDoLists(event)` to be an aysnc function             Changed a function `handleToggleToDoLists(event)` to be an aysnc function
            In an async function `handleRemoveToDoLists(event)`
            - Added an if statement to check `disabledItems.includes(id)` to return
            - Assigned `disabledItems = [...disabledItems, id]`
            - Added await to fetch
            - Re-assigned `disabledItems = disabledItems.filter((itemID) => itemID !== id)`

            In an async function `handleToggleToDoLists(event)`
            - Declared a const variable `id` with assigned value `event.detail.id`
            - Declared a const variable `value` with assigned value `event.detail.value`
            - Added an if statement to check `disabledItems.includes(id)` to return
            - Added await fetch with a PATCH request
            - Added `.then` to get an async response
            - Added an if...else statement to check the response
            - Defined a const variable `updateToDoLists` to store the response
            - Encapsulared `toDoLists` in an if's body of the statement
            - Re-assigned `disabledItems = disabledItems.filter((itemID) => itemID !== id)`

            Added a prop `{disabledItems}` to <ToDoList> component

Modified:   Defined a let variable `disabledItems` with a value of an empty array (src/lib/ToDoList.svelte)

            In an element `<li class:completed>`
            Added `disabled={disabledItems.includes(id)}` in <input> element
            Added `disabled={disabledItems.includes(id)}` in <button> element

            Added a property `&:disabled` in .remove-todo-button in a <style> block

+++ src/App.svelte
@@ -10,6 +10,7 @@
 	let error = null;
 	let isLoading = false;
 	let isAdding = false;
+	let disabledItems = [];
 
 	onMount(() => {
 		LOAD_toDoLists();
@@ -57,10 +58,13 @@
 		toDoListBox.focusInput();
 	}
 
-	function handleRemoveToDoLists(event) {
+	async function handleRemoveToDoLists(event) {
 		const id = event.detail.id;
+		if (disabledItems.includes(id)) return;
 
-		fetch(`https://jsonplaceholder.typicode.com/todos/${id}`, {
+		disabledItems = [...disabledItems, id];
+
+		await fetch(`https://jsonplaceholder.typicode.com/todos/${id}`, {
 			method: 'DELETE'
 		}).then((response) => {
 			if (response.ok) {
@@ -69,14 +73,36 @@
 				alert('An error has occured.');
 			}
 		});
+
+		disabledItems = disabledItems.filter((itemID) => itemID !== id);
 	}
 
-	function handleToggleToDoLists(event) {
-		toDoLists = toDoLists.map((toDoItem) => {
-			if (toDoItem.id === event.detail.id) {
-				return { ...toDoItem, completed: event.detail.value };
+	async function handleToggleToDoLists(event) {
+		const id = event.detail.id;
+		const value = event.detail.value;
+		if (disabledItems.includes(id)) return;
+
+		await fetch(`https://jsonplaceholder.typicode.com/todos/${id}`, {
+			method: 'PATCH',
+			body: JSON.stringify({
+				completed: value
+			}),
+			headers: {
+				'Content-type': 'application/json;charset=UTF-8'
 			}
-			return { ...toDoItem };
+		}).then(async (response) => {
+			if (response.ok) {
+				const updateToDoLists = await response.json();
+
+				toDoLists = toDoLists.map((toDoItem) => {
+					if (toDoItem.id === event.detail.id) {
+						return { ...toDoItem, completed: event.detail.value };
+					}
+					return { ...toDoItem };
+				});
+			}
+
+			disabledItems = disabledItems.filter((itemID) => itemID !== id);
 		});
 	}
 </script>
@@ -92,6 +118,7 @@
 			{toDoLists}
 			{error}
 			{isLoading}
+			{disabledItems}
 			disableAdding={isAdding}
 			bind:this={toDoListBox}
 			on:addtodo={handleAddToDoLists}
